+++ src/util.rs
@@ -6,7 +6,7 @@ use nalgebra::Vector3;
 use rand::prelude::{Rng, Distribution};
 use rand::distributions::Uniform;
 
-pub type Vec3 = Vector3<f32>;
+pub type Vec3 = Vector3<f64>;
 pub type Color = Vec3;
 pub type Point3 = Vec3;
 
@@ -25,7 +25,7 @@ impl AsRgb for Color {
     }
 
     fn as_rgb_multisample(self, samples_per_pixel: u32) -> Rgb<u8> {
-        let scale = 1.0 / samples_per_pixel as f32;
+        let scale = 1.0 / samples_per_pixel as f64;
         Rgb([
             ((self.x * scale).clamp(0.0, 0.999) * 256.0) as u8,
             ((self.y * scale).clamp(0.0, 0.999) * 256.0) as u8,
@@ -34,7 +34,7 @@ impl AsRgb for Color {
     }
 }
 
-pub fn vec3_random<D: Distribution<f32>, R: Rng>(distr: &D, rng: &mut R) -> Vec3 {
+pub fn vec3_random<D: Distribution<f64>, R: Rng>(distr: &D, rng: &mut R) -> Vec3 {
     Vec3::new(
         distr.sample(rng),
         distr.sample(rng),
@@ -43,7 +43,7 @@ pub fn vec3_random<D: Distribution<f32>, R: Rng>(distr: &D, rng: &mut R) -> Vec3
 }
 
 pub fn random_in_unit_sphere<R: Rng>(rng: &mut R) -> Vec3 {
-    let dist_m1p1 : Uniform<f32> = Uniform::new(-1.0, 1.0);
+    let dist_m1p1 : Uniform<f64> = Uniform::new(-1.0, 1.0);
     loop {
         let p = vec3_random(&dist_m1p1, rng);
         if p.magnitude_squared() < 1. {
@@ -62,7 +62,7 @@ impl Ray {
         Ray { orig, dir }
     }
 
-    pub fn at(&self, t: f32) -> Point3 {
+    pub fn at(&self, t: f64) -> Point3 {
         self.orig + self.dir * t
     }
